tests - refactoring java config
