+++ src/test/java/by/library/yavlash/config/ApplicationContextConfigurationTest.java
@@ -1,0 +1,102 @@
+package by.library.yavlash.config;
+
+import org.flywaydb.core.Flyway;
+import org.hibernate.jpa.HibernatePersistenceProvider;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.jdbc.datasource.DriverManagerDataSource;
+import org.springframework.orm.jpa.JpaTransactionManager;
+import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
+import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
+
+import javax.sql.DataSource;
+import java.util.Properties;
+
+@Configuration
+@ComponentScan("by.library.yavlash.repository")
+@PropertySource("classpath:/application.properties")
+@EnableJpaRepositories("by.library.yavlash.repository")
+public class ApplicationContextConfigurationTest {
+    @Value("${database.url}")
+    private String url;
+    @Value("${database.username}")
+    private String user;
+    @Value("${database.password}")
+    private String password;
+    @Value("${database.migration.location}")
+    private String migrationLocation;
+    @Value("${entity.path}")
+    private String entityPath;
+    @Value("${database.driver}")
+    private String driver;
+    @Value("${hibernate.dialect}")
+    private String dialect;
+    @Value("${hibernate.show_sql}")
+    private String showSql;
+    @Value("${hibernate.format_sql}")
+    private String formatSql;
+
+    @Bean(initMethod = "migrate")
+    public Flyway flyway() {
+        return Flyway.configure()
+                .dataSource(url, user, password)
+                .locations(migrationLocation)
+                .load();
+    }
+
+//    @Bean
+//    public LocalSessionFactoryBean sessionFactory() {
+//        LocalSessionFactoryBean sessionFactory = new LocalSessionFactoryBean();
+//        sessionFactory.setDataSource(dataSource());
+//        sessionFactory.setPackagesToScan(entityPath);
+//        sessionFactory.setHibernateProperties(hibernateProperties());
+//        return sessionFactory;
+//    }
+
+    @Bean
+    public DataSource dataSource() {
+        DriverManagerDataSource dataSource = new DriverManagerDataSource();
+        dataSource.setDriverClassName(driver);
+        dataSource.setUrl(url);
+        dataSource.setUsername(user);
+        dataSource.setPassword(password);
+        return dataSource;
+    }
+
+//        @Bean
+//    public Properties hibernateProperties() {
+//        Properties properties= new Properties();
+//        properties.setProperty("hibernate.connection.url", url);
+//        properties.setProperty("hibernate.connection.driver_class", driver);
+//        properties.setProperty("hibernate.connection.username", user);
+//        properties.setProperty("hibernate.connection.password", password);
+//        properties.setProperty("dialect", dialect);
+//        properties.setProperty("show_sql", showSql);
+//        properties.setProperty("format_sql", formatSql);
+//        return properties;
+//    }
+
+    @Bean
+    public LocalContainerEntityManagerFactoryBean entityManagerFactory() {
+        LocalContainerEntityManagerFactoryBean entityManagerFactoryBean = new LocalContainerEntityManagerFactoryBean();
+        entityManagerFactoryBean.setDataSource(dataSource());
+        entityManagerFactoryBean.setPersistenceProviderClass(HibernatePersistenceProvider.class);
+        entityManagerFactoryBean.setJpaVendorAdapter(new HibernateJpaVendorAdapter());
+        Properties jpaProperties = new Properties();
+        jpaProperties.setProperty("hibernate.dialect", "org.hibernate.dialect.DerbyTenSevenDialect");
+        entityManagerFactoryBean.setJpaProperties(jpaProperties);
+        entityManagerFactoryBean.setPackagesToScan(entityPath);
+        return entityManagerFactoryBean;
+    }
+
+    @Bean
+    public JpaTransactionManager transactionManager() {
+        JpaTransactionManager transactionManager = new JpaTransactionManager();
+        transactionManager.setEntityManagerFactory(entityManagerFactory().getObject());
+        return transactionManager;
+    }
+}
