--- src/test/java/by/library/yavlash/config/TestApplicationContextConfiguration.java
@@ -1,77 +1,0 @@
-package by.library.yavlash.config;
-
-import org.flywaydb.core.Flyway;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
-import org.springframework.orm.hibernate5.LocalSessionFactoryBean;
-
-import javax.sql.DataSource;
-import java.util.Properties;
-
-@Configuration
-@ComponentScan("by.library.yavlash.repository")
-@PropertySource("classpath:/application.properties")
-public class TestApplicationContextConfiguration {
-    @Value("${database.url}")
-    private String url;
-    @Value("${database.username}")
-    private String user;
-    @Value("${database.password}")
-    private String password;
-    @Value("${database.migration.location}")
-    private String migrationLocation;
-    @Value("${entity.path}")
-    private String entityPath;
-    @Value("${database.driver}")
-    private String driver;
-    @Value("${hibernate.dialect}")
-    private String dialect;
-    @Value("${hibernate.show_sql}")
-    private String showSql;
-    @Value("${hibernate.format_sql}")
-    private String formatSql;
-
-    @Bean(initMethod = "migrate")
-    public Flyway flyway() {
-        return Flyway.configure()
-                .dataSource(url, user, password)
-                .locations(migrationLocation)
-                .load();
-    }
-
-    @Bean
-    public LocalSessionFactoryBean sessionFactory() {
-        LocalSessionFactoryBean sessionFactory = new LocalSessionFactoryBean();
-        sessionFactory.setDataSource(dataSource());
-        sessionFactory.setPackagesToScan(entityPath);
-        sessionFactory.setHibernateProperties(hibernateProperties());
-        return sessionFactory;
-    }
-
-    @Bean
-    public DataSource dataSource() {
-        DriverManagerDataSource dataSource = new DriverManagerDataSource();
-        dataSource.setDriverClassName(driver);
-        dataSource.setUrl(url);
-        dataSource.setUsername(user);
-        dataSource.setPassword(password);
-        return dataSource;
-    }
-
-    @Bean
-    public Properties hibernateProperties() {
-        Properties properties= new Properties();
-        properties.setProperty("hibernate.connection.url", url);
-        properties.setProperty("hibernate.connection.driver_class", driver);
-        properties.setProperty("hibernate.connection.username", user);
-        properties.setProperty("hibernate.connection.password", password);
-        properties.setProperty("dialect", dialect);
-        properties.setProperty("show_sql", showSql);
-        properties.setProperty("format_sql", formatSql);
-        return properties;
-    }
-}
