--- src/test/java/by/library/yavlash/config/TestServiceConfiguration.java
@@ -1,9 +1,0 @@
-package by.library.yavlash.config;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ComponentScan("by.library.yavlash.service")
-public class TestServiceConfiguration {
-}
